+++ Sources/PatchDay/Hormones/Cell/HormoneCellViewModel.swift
@@ -16,6 +16,8 @@ class HormoneCellViewModel: HormoneCellViewModelProtocol {
     private let sdk: PatchDataSDK
     private let isPad: Bool
 
+    var now: PDNow = PDNow()
+
     required init(cellIndex: Index, sdk: PatchDataSDK, isPad: Bool) {
         self.cellIndex = cellIndex
         self.sdk = sdk
@@ -69,7 +71,11 @@ class HormoneCellViewModel: HormoneCellViewModelProtocol {
         guard let expiration = hormone.expiration else { return nil }
         guard !hormone.date.isDefault() else { return nil }
         let prefix = HormoneStrings.create(hormone).expirationText
-        let dateString = PDDateFormatter.formatDay(expiration)
+
+        let dateString = expiration.isWithin(minutes: Minutes.IN_WEEK, of: now.now)
+            ? PDDateFormatter.formatDay(expiration)
+            : PDDateFormatter.formatDate(expiration)
+
         return "\(prefix) \(dateString)"
     }