+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Fixed
 
 Changed
 
+- Hormone Cells now display the full expiration date when the hormone is a week or more away from being
+    expired or if it has been a week or more since the hormone expired.
 - Pill Cells now don't display a Last Taken date when the pill was not yet taken that day.
 - You can now save typed Pill name changes without having to close the text field.
 - You can now save typed Site name changes without having to close the text field.